requirements.txt
streamlit==1.46.1
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-groq==0.3.4
langchain-huggingface==0.3.0
faiss-cpu==1.11.0
sentence-transformers==4.1.0
huggingface-hub==0.33.1
python-dotenv==1.1.1